Add LichBullet

관련 이슈 : #47

+++ Assets/Scripts/MinionScripts/MinionStat.cs
@@ -4,15 +4,15 @@ using UnityEngine;
 
 public class MinionStat : MonoBehaviour
 {
-    [SerializeField] int _maxHP;
-    [SerializeField] int _hp;
-    [SerializeField] int _atk;
-    [SerializeField] int _def;
+    [SerializeField] float _maxHP;
+    [SerializeField] float _hp;
+    [SerializeField] float _atk;
+    [SerializeField] float _def;
 
-    public int MaxHP { get { return _maxHP; } set { _maxHP = value; } }
-    public int HP { get { return _hp; } set { _hp = value; } }
-    public int Atk { get { return _atk; } set { _atk = value; } }
-    public int Def { get { return _def; } set { _def = value; } }
+    public float MaxHP { get { return _maxHP; } set { _maxHP = value; } }
+    public float HP { get { return _hp; } set { _hp = value; } }
+    public float Atk { get { return _atk; } set { _atk = value; } }
+    public float Def { get { return _def; } set { _def = value; } }
 
     private void Start()
     {
